DATAJPA-696 - More robust support for nested subgraphs.
Moved PersistenceProvider agnostic functionality to Jpa21Utils.
